row with different background, entry info
Folder icon and music icon is added to the row
different background for folder and different for file
listview selector state added
More formating for showing title/album view and lables
Entry info is also shown now. For directory - no. of entries in it
and for file - file size in proper format